Support property getters and setters and properly implement prototype inheritance. Getters and setters are now assigned using "get" and "set" fields in the property declaration. They can't and shouldn't be set from the main script using the previous this.properties.<property-name>.setter/getter notation. getter and setter functions aren't visible to the JavaScript on the node and are fixed once declared.
Properties may be initialized using either a simple initializer (as before), or using a hash-based initializer that may contain "get", "set", and "value" fields. "value" specifies the initial value of the property. "get" and "set" specify the body of the accessor functions.  The getter should execute a return statement to provide the outgoing value. The incoming value for the setter is provided by an "value" variable automatically provided. Accessor functions may read and write backing storage for the property using this.properties.<property-name> or this.<property-name>. The system knows when a property reference is made within one of that property's accessor functions and will avoid the infinite recursion.

Values on prototype objects will be visible on derived objects. Accessor functions on prototype objects will execute for derived objects even after the property on the derived object has been overwritten with its own value.


git-svn-id: http://svn.virtualworldframework.com/trunk@297 775f9103-8eee-4f94-88be-65e0ca6d8b75