Suppress model API functions in stages for functions that aren't implemented in the stage to the right. Calls shouldn't be passed all the way through the pipeline just to be dropped at the actual driver.
Rework the construction process so that stages receive a callback as each is connected, first from right-to-left (modelize), and then from left-to-right (kernelize). Model API suppression needs to happen in the right-to-left phase, and it wasn't possible with the previous construction which only called initialize() left-to-right.

Add new vwf/api/kernel and vwf/api/model modules to declare the kernel and model APIs, respectively. They are currently used just for the function names for the API generators and model API suppression, but can be extended to declare parameters and allow many of the kernel and driver functions to be generated automatically.


git-svn-id: http://svn.virtualworldframework.com/trunk@358 775f9103-8eee-4f94-88be-65e0ca6d8b75