Interpret a JavaScript handler that doesn't return anything (and implicitly returns undefined) as returning true. This allows the simplest implementation in the handlers for the common cases of:
  - Adding a handler to block downward propagation in the capture phase, or
  - Handling and absorbing an event in the bubble phase

A handler should return a truthy value or nothing to cancel propagation and a falsy value other than undefined to allow propagation to continue. The latter case is the minority case.

-----

Invoke a JavaScript handler that didn't provide a context on the global root node. This is similar to JavaScript's behavior where "this" defaults the JavaScript global object.

Invoke a JavaScript handler added directly (not wrapped in the add/handler/context object provided by the events.add() helper) in the context of the source. This is useful for dispatched events, which are intended to be handled by the originating node.

  - node.*event* = this.events.add( function() { ... }, this ); // normal case, explicit context: invoked on this, or whichever node is specified
  - node.*event* = this.events.add( function() { ... } ); // no context: invoked on global root
  - this.*click* = function() { ... }; // implicit context: invoked on this, or whichever node is the event source


git-svn-id: http://svn.virtualworldframework.com/branches/integration@691 775f9103-8eee-4f94-88be-65e0ca6d8b75