Add initializingProperty()/initializedProperty() driver functions and distinguish between creating a new property on a node, fully suppressing a property of the same name on a prototype, and setting the value the first time for a property defined on a prototype. kernel.createProperty() always calls into the the drivers with creating/created, as before. kernel.setProperty() calls into the drivers with initializing/initialized if the property is not defined directly only the node yet, and with setting/sat otherwise, as before.
kernel.createNode() initializes new properties with kernel.createProperty() when the specification contains accessor functions or when the property is not defined on any prototype, and with kernel.setProperty() otherwise to initialize the new node's copy of the property. kernel.setProperty() calls into the drivers using initializing/initialized in this case.

Since properties may provide accessor functions, but also use the value store, it's necessary to make this distinction to allow derived nodes to properly initialize inherited properties.


git-svn-id: http://svn.virtualworldframework.com/branches/integration@886 775f9103-8eee-4f94-88be-65e0ca6d8b75