Merge the kernel's logger() and logger_for() together and move into a RequireJS module.
Use this.logger.log( ... ), etc. for normal use when there's just a message or a set of objects. The message will have the context string prepended.

Use this.logger.logc( context, ... ), etc. to add *context* as a one-time extra piece of context (such as the name of the current function). This was the interface formerly created for the drivers (using logger_for()) that they called as this.logger.log( context, ...), etc.

Now, log( ... ), etc. always handler the normal case, and logc( context, ... ) is for the case when extra context is provided.

The JavaScript driver now provides a logger to each node as this.logger.log(), etc.

logger_for() has been removed from the kernel, however its logger() has to remain until the kernel is converted to a RequireJS module and can load after the logger module.


git-svn-id: http://svn.virtualworldframework.com/branches/integration@984 775f9103-8eee-4f94-88be-65e0ca6d8b75