Clean up undefined kernel and model/view API references that were appearing in the vwf/view/document (vwf_view) prototype--and potentially in other drivers as well.
vwf/view/document provides kernel and view function generators but it doesn't generate any functions. These were being explicitly assigned as undefined values, but to avoid confusion they should be filtered out. The kernel functions are ones that would be called by an upstream stage or driver if this were a pipeline stage. They aren't applicable for a driver at the end of the pipeline. The model/view functions are ones that would be called by a downstream stage or the kernel. If the driver doesn't require them, then they may be either undefined or omitted with the same effect.

The references to undefined kernel functions in particular were causing confusion. Since drivers need to call downstream through the "kernel" property, and not through the prototype, having ghosts of the kernel functions on the prototype was misleading.


git-svn-id: http://svn.virtualworldframework.com/vwf/development@2017 775f9103-8eee-4f94-88be-65e0ca6d8b75