Move the queue suspend/resume down around the sections that actually perform the async calls. There are only two: jQuery.ajax() in loadComponent(), and the creatingNode()/createdNode() driver calls in createChild().
Previously, we were wrapping all the functions that ultimately called loadComponent() or createNode() as they were pulled from the queue and executed. That had the advantage of pausing and resuming the queue in just one place, and doing this near dispatch() where other queue work occurred.

But that wasn't sufficient since those actions could be called during the execution of many other actions. To prevent actions from executing out of order, queue processing must be suspended while any async operation is in progress within the model, regardless of the origin.

There are two important points about this change:

  - If a new async operation is added in the future, it must also be wrapped with suspend/resume calls. There is no structural way to enforce this, but comments in multiple places near the async operations hope to make this clear.

  - queue.resume() invokes dispatch() to resume execution when the queue turns back on. Since queue.resume() is invoked by one action as it completes, and since dispatch() may execute other actions, any functions that call queue.resume() must be sure to do so immediately before returning to the host (the browser). Otherwise, work from the completing action may occur after work from a new action that dispatch() starts. Comments around the existing calls to queue.resume() try to make this requirement clear as well.


git-svn-id: http://svn.virtualworldframework.com/vwf/branches/id@2159 775f9103-8eee-4f94-88be-65e0ca6d8b75