Create a property when kernel.setProperty() is called for a property that doesn't exist on the node or on one of its prototypes.
Previously, kernel.setProperty() would set the property if it already existed on the node, and initialize the property if it existed on a prototype but not on the node. If it didn't exist on either the node or a prototype, kernel.setProperty() would initialize the property by default.

However, it's incorrect to initialize a property that doesn't exist. The alternatives are to either do nothing, and require that the property first be created before it can be set, or to create the property automatically. Creating the property automatically is the more JavaScript-like option, and is appropriate here.


git-svn-id: http://svn.virtualworldframework.com/vwf/branches/id@2334 775f9103-8eee-4f94-88be-65e0ca6d8b75