Make listeners that are attached to behaviors work again.
Unlike property getters and setters and method bodies, event listeners may be attached above the node the event is defined on. The behavior proxy wasn't getting a copy of event listeners if the corresponding event wasn't defined directly on the behavior.

Before r2330, this was working since the listener search (incorrectly) followed the prototype chain directly to the behavior and its prototypes. Now that the prototype chain passes through the behavior proxy, and not the actual behavior, some listeners defined on the behavior were being missed.

Fixes the duck's spin-on-click behavior, and probably other undiscovered problems.


git-svn-id: http://svn.virtualworldframework.com/vwf/branches/id@2390 775f9103-8eee-4f94-88be-65e0ca6d8b75