Remove `origin` from the behavior `proxy.private`.
Self-targeted events are inherited from prototypes, such as for:

    this.pointerClick = function( ... ) { ... }

A search for listeners for these events needs to identify those that
target the same node that they are attached to. But when a node
`implement`s a behavior, the node in the prototype chain is a proxy to
the actual behavior node. The search needs to recognize listeners on the
proxy as self-targeting.

Previously, the proxy contained an `origin` field that referred to the
behavior, and the listener search tested against both the node (the
proxy) and the node's `origin`.

This change removes the `origin` field and instead remaps the proxy's
listeners to target the proxy instead of the behavior. The listener
search can then simply test against the node.

This is in preparation for removing the driver's `private` bookkeeping
field from the node.