Block kernel reentry from view drivers during replication.
While the application state is being read or written during replication,
kernel calls from model drivers are blocked to prevent scripts from
coloring the results. Only the underlying data is relevant.  Property
getters and setters, and initialization scripts could cause the state to
change as it is read, or modify it when it is written.

View drivers can't affect the model state, so they don't need to be
blocked during replication for correctness. But, since they do see the
same `getProperty`, `setProperty`, `createChild`, etc. calls that the
models see, an application could call the kernel during replication and
create unwanted reflector messages.