REFACTOR: Add immediate, read-only `kernel.property` for views.
Allows views to read value properties and the most recently set values
of accessor properties without waiting for a reflector hop. `vwf.find`
and `vwf.getProperty` were the most common reasons for view drivers and
application html to want to use the deprecated `vwf` global.

`kernel.property` can be used in place of `kernel.getProperty`.
`kernel.find` uses the same data as `kernel.property`, and it is
safe for views to use too.

In `setProperty`, call `model.settingProperty` on all of the drivers,
even after the value has been "handled". All of the drivers need to be
able to monitor property changes, and model/object needs to see them to
set the change flag and to record values for `kernel.property` to use.

However, don't call `model.settingProperty` for accessor properties
after the accessor function has delegated to whereever it delegates to.
Purely delegated properties such as the `node3` transform components
were being recorded in model/object as changed even though they don't
have a primary value.