Remove the Sugar libary from the kernel.
Sugar is good for the application environment where it's carefully
controlled. But we should be more conservative at the system level since
there are many more third-party libraries interacting. Underscore or
Lo-Dash are better choices when we get to it.

Undoes the fix in commit 8dd8492 since it used a Sugar function.

Closes #1520. References #1622.