Replace `vwf_view.kernel` with `this.kernel`.
`vwf_view` is a global created by `view/document` to support global code
in the `application.vwf.html` file. It refers explicitly to the
`vwf/document` driver, and it won't necessarily exist since
`vwf/document` could be omitted.

Within the editor, `this` refers to `view/editor` and is the correct
reference for calling kernel actions from the editor.