Replace `childURI` with `childIndex` and expand initializingNode.
For global nodes, pass the URI as the childIndex/childURI argument as
before. But for descendant nodes, pass the index number of the child's
position in the parent's array instead. Nodes can load out of order, so
the drivers need to know where to attach the new node.

Also, provide the `creating`/`createdNode` arguments to
`initializingNode`/`initializedNode`. Drivers that only need the
`childIndex` or other parameters when the node finishes construction
won't need to store them when it starts constructing.

References #1918.