Keep child array order consistent with the child list in the component.
Previously, when creating child nodes from a component, only the
name-based associations were consistent. Child nodes load in an
unpredictable order, so the number-based associations could describe a
different order from that of the component. The order could be different
on different clients as well.

For example, for the component:

  ---
  children:
    alpha: ...
    beta: ...

`node.children.alpha` and `node.children.beta` always referred to the
correct nodes, but node.children[0] and node.children[1] could refer to
either `beta` or `alpha`, and that association could be different from
client-to-client and run-to-run.

This commit moves initial registration of the child node ahead of
where its prototype and behavior dependencies are resolved so that
children appear at the parent before they perform any async operations
and in a consistent order. Along with the childURI/childIndex change,
drivers can now maintain a child array in the correct order.

The server must also maintain the child order when loading components
and converting to JSON, but this is now the case since Ruby 1.9.3 is the
minimum requirement.

Closes #1918.