Don't remove reflector messages in `setState` sent after the action.
When updating the message queue to match the incoming specification,
`setState` previously cleared the queue except for messages with a
`respond` field. But that would delete messages sent after the
`setState` that arrived before it finished executing. Those messages
apply to the new state and should be retained.

The correct process is to delete all future messages, including any
generated during the `setState` since the incoming queue contains the
correct messages for the new state, delete any reflector messages that
arrived before the `setState` but were scheduled to execute at a later
time, but retain any reflector messages that arrived after the
`setState`.

Fixes #2122.