Re: Redmine Task #2196
Update for minor appearance tweaks.
Date format was already handled when the
datestamp/username 'column' was added.

Turned off text area resizing for the chat input
box.

Added HTML escaping functionality to the user messaging
system, so posting HTML to the chat should display properly
and not be interpreted as HTML.
A