Create a global `clients.vwf` node and maintain one child per client.
The server creates the global node `http://vwf.example.com/clients.vwf`
on launch (independent of the application) then adds one child to it for
each client connected to the application instance. The child's name is
the client's moniker.

In the default implementation, neither the `clients.vwf` node nor its
children contain any additional data. It is expected that server
installations that have user data to share will make that available to
the application by creating the nodes with additional behaviors and
properties.

References #1272.