Begin moving the node registry to the kernel from model/object.
The kernel has been delegating knowledge of the node graph and node
change flags to the model/object driver. But this doesn't work well
since model/object isn't normally notified about changes to a node until
after the other drivers. Extra calls into model/object are in place so
that preceding drivers can ask about a node's ancestors or prototypes in
creatingNode, for example, but this is an abuse of the driver API.

Tracking changes in model/object also requires extra work since
model/object doesn't see changes to properties handled by other drivers.
Change flags may not be set in some cases, which can cause replication
errors.

The kernel will be better off if it manages this data itself. With a
local registry, it will be easier for the kernel to register a new node
just before it is first exposed to any driver and unregister a deleted
node right after its last driver message.

This commit creates the node and per-node property registries, and
records per-node property change flags.

Fixes #1815.