Track kernel re-entry originating with aggregate property sets and gets.
Set the markers so that if `kernel.setProperties` or
`kernel.getProperties` set or get a property having an accessor function
that refers back to the property, or if a driver calls back to the
kernel to refer to the property, `setProperty` or `getProperty` will
start the action with the correct driver.

Previously, only actions originating with `setProperty` or `getProperty`
were tracked this way. This commit adds tracking for actions originating
in `setProperties` and `getProperties` as well.