Allow descriptors to declare properties without assigning a value.
`kernel.createProperty( nodeID, name, undefined )` will create a new
property without assigning an initial value. This can be useful if the
property should have the value `undefined` or if a driver binds the node
to an internal object that already has a value (such as `transform` in a
Collada object).

But since YAML and JSON have no representation for `undefined`, there
has been no way to do the same thing using a component descriptor.

This commit extends the detailed initializer format slightly to allow an
`undefined` value to be explicitly declared:

    properties:
	  alpha:
	    "alpha"
	  beta:
	    value: "beta"
	  gamma:
	    undefined: true

`alpha` is created with the value `"alpha"` using a simple initializer.
`beta` is created with the value `"beta"` using a detailed initializer.
And `gamma` is created using a detailed initializer with the initial
value undefined.