Move the `transforms.transit` normalization just before the socket send.
Parameter and result values need to be normalized before they are
converted to JSON and sent to the reflector. Specifically, Array-like
objects must be converted to Arrays so that they pass through JSON
serialization and deserialization properly.

Actions initiated by views and sent using `kernel.send` weren't being
normalized this way.

This commit moves normalization into `kernel.send` and `kernel.respond`
to ensure that it is applied consistently for each outgoing message.
Other `transforms.transit` normalizations done in anticipation of the
result being sent to the reflector are redundant and have been removed.

The patch to explicitly convert array-likes to arrays in view/threejs is
now also redundant and has been removed.

Fixes #2666.