Load the application chrome as a service instead of a file.
Instead of having the client parse the URL and guess at the application
name, load the `application.vwf.html` overlay using a fixed path
relative to the application root.

Since applications aren't required to provide an overlay, return an
empty result with status code 200 when the overlay doesn't exist.
Otherwise, the log will show an error for a condition that isn't really
an error.

Closes #741.