Wait for the application node to load before initializing the view.
The HTML overlay previously loaded whenever it was ready, some time
after starting construction of the application tree. Some child nodes
may have already been contructed by that point, depending on timing and
whether or not they were contained in the application's component and
had other dependencies.

As of commit:12c139f, the application waits for the overlay to load
before continuing past the application root.

The tile puzzle was expecting certain child nodes to exist immediately
after the overload loaded. The fix is to wait for the application to
load (to receive `initiaizedNode`) before interacting with it.

Fixes #2720.