Add first support for a device to notify the reflector that it is a device
Right now the reflector recognizes that client as a device and does not send any messages to it.  Eventually, it will only send messages that pertain to the component that the device client instantiated.  There is still work to be done to deal with the case of a device client being the first to connect to an application instance and to deal w/ getState and setState