Adds in a fake authoritative state to both the nodeJS and ruby server implementations.  Fake state is managed by requesting state from the first client if no getState has been fullfilled in the last ten seconds, and storing the state from any getState request made.
Adds test function for determining if persistence is enabled for
any specific application.

Adds functionality to reflector where instances are marked as
being enabled for persistence, and loading previous persistence
state if one is present and persistence is enabled for the application.

Adds functionality to save the authoritative state when the last
client attached to an instance disconnects.

References #2768.
References #2769.
References #2770.
References #2771.
References #2737.