Turn off view interpolation on the camera node (previously was only navObject)
When a user controls the navObject in walk mode, the yaw is applied to the navObject and the pitch is applied to the camera.  Therefore, both objects (if they are different objects) are under user control and the latency avoidance applied to objects under user control conflicts with view interpolation, so we must turn it off.