Don't include messages following the current action in `kernel.getState.
In the queue returned by `kernel.getState`, exclude reflector messages
sent after the currently-executing action. These messages represent
external actions from the future and should not be included in the
replication state.

This is complementary with the queue handling in `kernel.setState`
which includes messages received after the currently-executing action
with actions from the received queue.