Partial fix for `node.create( ... )` followed by `node.children()`.
`node.children()` is ambiguous between the time of calling `node.create()`
and the end of the action, when the create finally resolves. The new node
should be hidden from outer tree while it is under construction, but
currently it is not. Calling `node.find()` from the parent node or higher
with a recursive query will attempt to search into the constructing node.
This triggers an exceptions with the "does not have a valid behaviors
array" message.

This commit hides the child node properly and fixes the `find` exception.
How child nodes under construction should (or should not) be represented
in `node.children` still needs to be resolved.

References #3789.