Clarify names that reference the proto-prototype node `node.vwf`.
Rename the `kernel/utility` items `nodeTypeURI` and `nodeTypeDescriptor`
to `protoNodeURI` and `protoNodeDescriptor`, respectively, to better
reflect that they refer to the proto-prototype node.

Rename the `model/javascript` internal variable `nodeNode` to
`protoNodeNode` so that it parallels the `protoNodeURI` name it derives
from and to hopefully clarify that it is the `model/javascript`
`...Node` data for the `protoNode...` node.

References #3088.