Add a property to control the animation tick rate.
It may be useful to reduce the rate at which animations are calculated
to reduce the CPU load when 60 fps animations are not required.

The tick rate may be set globally for all nodes that implement
`animation.vwf` (after the animation component has loaded):

    node.find( "doc('http://vwf.example.com/animation.vwf')" )[0].animationTPS = 30;

Or just for individual nodes:

    node.animationTPS = 10;

The default animation tick rate is unchanged at 60 per second.