Remove instance delegating to latest revision.
Move toward keeping separate states and actions for
automatically-recorded data and using revisions just for the tags and an
indication of a user-created snapshot.

States and actions can be combined to form revision states at any point,
and internal state and action list can be more easily flushed when no
revision refers to them.