Patch getIntersection to limit search after finding edge pixels
When `getIntersection` encounters an antialiasing pixel in the hit graph,
it can't match it to an object since the color is off. But since it knows
an object is near, it continues the search at a slightly larger radius.

However, once it decides to search the larger distances, the search only
terminates if an object is found. Unsuccessful search passes are ignored,
and the area can expand to find objects at an arbitrary distance away. If
no object is found before the search expands past the edges of the canvas,
the search will never complete.

This patch aborts the search once an empty pixel is discovered. Since
areas outside the canvas are treated as empty, at worst the search will
complete once it reaches an edge. This change is present in later versions
of Konva. Updating this library to one of those versions won't cause the
bug to return.

This change causes the seach to be slightly more conservative than
necessary. Ideally, it would complete the search at a particular radius
before giving up rather than stopping once the first empty pixel is found.
Some objects that are just one pixel from the reference point may be
missed, depending on the relative order of empty, edge, and occupied
pixels in the search order. However, the improvement isn't worth making
this historical version incompatible with the implementation in current
versions. The change should be applied upstream then backported if it's
needed.