VWF: Manage clients.vwf children and pending messages correctly
Resetting the state for all clients each time a new client joins was
not incorporating messages that occurred after the state was snapped for
exising clients. This includes the message to create the client's child in
clients.vwf as well as other client messages that arrive while waiting for
the `getState` result.

New clients would correctly receive the initial state and the pending
messages. Existing clients would receive the clients.vwf message and other
intermediate messages as they occurred, then receive the `setState`. This
would rewind the state to before the clients.vwf and intermediate messages
leaving existing clients in the wrong state and in a different state from
the new client.