Correct `mouseDown` when the button changes away from a listening node
Konva `mousedown`/`mouseup` listeners are frequently only active on a
subset of the nodes and not across the entire canvas. If the mouse button
is pressed while over a node with an active `mousedown` listener and
released outside of any node without an active `mouseup` listener, we
never see a `mouseup` event and the `mouseDown` variable remains set.

Since `mouseDown` indicates that enter and exit events should become swipe
events, this causes spurious swipe events as the mouse is moved across the
drawing.