When user drags a node, set its mapPosition directly
Previously, it called the setMapPositionFromPosition method, but as a result, external applications receiving reflector traffic were never sent the new mapPosition since that was calculated internally in the model.  This is a little bit hacky as the code in the comment describes, but it's the best option for our use case.