Reduce timeout from 15s to 5s.
At 15s, the screen stays blank for about 15s on a lobby refresh.  That's not a good user experience.
Too long and a user is going to repeatedly hit the refresh button in frustration.
At a 5s timeout, the screen is blank for 1-5s, far more tolerable for a user and within the
recommended range.